Fixed everything from scratch. Now it works.
Although, the report is not very done. It needs some text explaining. Do you guys also think I should compare graphs? i.e. have a figure with both numerical and exact on it? Another question, how do we see which order the log-log plot is in? am a bit clapped.